requirements.txt
pytest==7.2.1
Appium-Python-Client==1.0.2
